backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
# asyncpg is the async PostgreSQL driver — psycopg2 not needed for async usage
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
greenlet>=3.0.0

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Authentication
# Using bcrypt directly — passlib 1.7.4 is incompatible with bcrypt >=4.x
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# HTTP Client (Paystack integration)
httpx>=0.25.0

# File handling
python-multipart>=0.0.6
aiofiles>=23.0.0

# Environment
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
